ticket_broker/requirements.txt
camunda-external-task-client-python3==4.3.0
certifi==2022.12.7
chardet==3.0.4
idna==2.10
pydantic==1.10.2
requests==2.24.0
typing_extensions==4.4.0
urllib3==1.25.11
networkx==2.8.8
pytest==7.2.0
pytest-cov==4.0.0
SQLAlchemy==1.4.4
PyYAML==6.0